requirements.txt
alembic==1.0.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
elasticsearch==6.3.1
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Dance==1.1.0
Flask-Excel==0.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
geographiclib==1.49
geopy==1.16.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lazy==1.3
lml==0.0.4
Mako==1.0.7
MarkupSafe==1.0
newrelic==4.4.0.103
numpy==1.15.0
oauth2client==4.1.3
oauthlib==2.1.0
openpyxl==2.5.6
pandas==0.23.4
passlib==1.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
rauth==0.7.3
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rq==0.12.0
rsa==4.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.10
SQLAlchemy-Utils==0.33.5
texttable==1.4.0
transitions==0.6.8
uritemplate==3.0.0
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.1.1
xlwt==1.3.0
